fix test for Go 1.15
